requirements.txt
streamlit
mysql-connector-python
pickle